openapi3: move originTree into origin.go
It is origin machinery, and marsh.go is expected to hold the only direct yaml
uses outside origin.go (the CI unmarshal gate pins that count).
