+++ openapi3/origin.go
@@ -1,6 +1,7 @@
 package openapi3
 
 import (
+	"net/url"
 	"reflect"
 	"sort"
 	"strings"
@@ -323,4 +324,25 @@ func jsonTagName(f reflect.StructField) string {
 	}
 	name, _, _ := strings.Cut(tag, ",")
 	return name
+}
+
+// originTree parses data solely for its origin tree, used to re-attach source
+// origins to a component resolved through the generic-map path: a $ref to a
+// schema under an arbitrary top-level key lands in T.Extensions and loses its
+// origin in the json round-trip (see resolveComponent). Returns nil when parsing
+// fails; callers degrade to no origin.
+func originTree(data []byte, location *url.URL) *yaml.OriginTree {
+	var file string
+	if location != nil {
+		file = location.String()
+	}
+	var v any
+	tree, err := yaml.Unmarshal(data, &v, yaml.DecodeOpts{
+		Origin:            yaml.OriginOpt{Enabled: true, File: file},
+		DisableTimestamps: true,
+	})
+	if err != nil {
+		return nil
+	}
+	return tree
 }
